LUA Functions Added, AttackAllowed, IsInRaid, InSameRaid, GetRaid
AttackAllowed(Spawn, Target, Distance, RangeAttack)
    Distance default is 0.0f and just checks if you can attack the target.
    Otherwise distance greater than 0.0f relates to the distance between the Spawn and Target which you first obtain using GetDistance(Spawn, Target).
    RangeAttack is a boolean (default: false) when true we use range info for the ranged attack min/max distance versus max distance melee combat range.

IsInRaid(Spawn, IsLeaderGroup) - Establishes if the Spawn is currently in a raid.    IsLeaderGroup is boolean (default: false) returns true if Spawn is the leader group of the raid.

InSameRaid(Spawn, Target, IsLeaderGroup) - determines if Spawn and Target are within the same raid groups.  IsLeaderGroup is boolean (default: false) returns true if Spawn is the leader group of the raid.

GetRaid(Spawn, PlayersOnly) - gets the array of all the members in the raid.  PlayersOnly is boolean (default: false), if true we return only clients/players in the list for the raid.